Add release command
It can be used either for specific artifact IDs or from a branch.

I've added a context.Context to our Git operations as well to handle timeouts and aborts from the client for long running actions.

Lastly have I added a skip option to the logger to make the "caller" field in logs be of our source code instead of the logging package's.