Add authenticate middleware
The added authenticate function takes a token that requests are
authenticated against.

Further more daemon and hamctl tokens are moved into server flags with
fallback to env variabels (current behaviour).

I moved the Options struct to the http NewServer function instead of
having it in the command packaged as it only specified HTTP options.
This simplifies calls to the NewServer function.