Align naaming on artifacts (#30)
We have used build, release, spec and others for the same thing. This
change aligns on the term artifact.
An artifact still has a specification, ie. artifact.Spec.

We also change to place where artifacts are stored from "builds/" to
"artifacts/" in the config repository.
This is a breaking change so shuttle plans must be updated as well.