Add push command to artifact CLI
This will commiit a directory as an artifact to a configuration
repository.

The test_push make target will generate an artifact specification and a
config map yaml file and push it to a config repo.

I moved the commit messages into a central file as well as we depend
on the format of these messages in the release flows.