Move service and git-branch flags to artifact init
We did not store the service name or origin branch information in the
artifact specification. This makes in unnecessarily hard to get these
when handling artifacts.

With this change the init command takes a service, git-branch and
optional shuttle-plan-branch flag to set this information.

This simplifies the push command to only take flags for the push
operation.