Add push command to artifact CLI (#31)
* Add push command to artifact CLI

This will commit a directory as an artifact to a configuration
repository.

The test_push make target will generate an artifact specification and a
config map yaml file and push it to a config repo.

I moved the commit messages into a central file as well as we depend
on the format of these messages in the release flows.

* Move service and git-branch flags to artifact init

We did not store the service name or origin branch information in the
artifact specification. This makes in unnecessarily hard to get these
when handling artifacts.

With this change the init command takes a service, git-branch and
optional shuttle-plan-branch flag to set this information.

This simplifies the push command to only take flags for the push
operation.