Validate user mapping input
If an empty user mapping was provided articat and server commands would
panic.

This change adds input validation and sanitization of the user-mappings
flag and environment variable.
Empty strings are filtered out and whitespaces are trimmed. Conflicting
mappings are detected as well as incomplete mappings.

A user-mappings flag is added to "artifact init" command to make the
configuration discoverable as well.