Harden commit message lookup functions (#54)
* Harden commit message lookup functions

When looking up releases, artifacts etc. we did it using
strings.Contains() calls. This made partial matches appear to be valid,
ie. release master-123 would match any artifact starting with that
string.

This change ensures that we match on the exact fields and at the same
ignores casing in the results.

* Align test setup between locate condition tests