Fix daemon notifications for default namespaces
If the daemon receives a webhook for a deployment in dev and the located
release is in another environment, we were unable to load the artifact
spec as the event's namespace is not matching the release.

This ensures that we default to env as namespace if it is one of the
environment namespaces dev, staging and prod.