Configurable logging with flags and environment variables (#74)
Add log.level and log.development flags to server and daemon commands and read values from environment if available from LOG_LEVEL and LOG_DEVELOPMENT respectively.

This makes is possible to configure the logging level and output format. If development is set, it is human readable with level, caller and message only.

Non-development is still JSON with all fields.