Add details to failed auto-release slack message
We are currently writing a single slack message in case of any error
during auto-release policies. This hides the details of what
auto-release policy actually failed, if there are multiple.

This change writes a slack message on each iteration on auto-release
policies. Further more does it align the message with the parameters.