Tag GitHub repositories on releases
This change adds a tagging feature for GitHub application repositories.
It is feature toggled by the github-api-token CLI flag.

If the token is provided and the released application is hosted on
GitHub, we attempt to add a tag named after the environment on the Git
sha.

With this change the main logic of flow.notifyRelease is refactored out
to a hook on the flow.Service stuct. This simplifies testing and
prepares for a more testable flow package.

Further more the notificatioin is now done in a Go routine. This
decreases request response times and we did not propagate any errors out
to the user anyway.