Support prerelease tags
This change deprecates the mutable docker tags latest vX and vX.Y. We do not use
these any ways and Goreleaser don't have conditionally selecting the docer image
templates.

This unlocks the posibility of pushing v1.2.3-alpha1 tags that gets build and
released to GitHub as prerelease and as Docker images with the complete tag
name.