Native copy command (#122)
As an attempt to optimize copy performance this change uses Linux cp command
instead of github.com/otiai10/copy to copy the master repository.

The module is responsible for most CPU work and allocations naturally as this is
one of the main operations of the release-manager. Hopefully by using a native
implementation the time spend on these operations is optimized.