Handle unknown configuration
If a client mistypes an environment, service or branch name we currently fail
with a 500 internal error. This is a bug introduced with commits 62439e7,
7abe09f and 738f3ce when we switched to copy mechanism from Go std lib to native
cp execs.

This change makes package copy expose an ErrUnknownSource error if the source
path is not found which error handling is updated to detect.