Handle ErrBranchBehindOrigin errors (#127)
On racing writes to the config repo ErrBranchBehindOrigin is returned. But as it
is not mapped in our HTTP handlers to a client facing message they just receive
an "unknown error" message.

This change maps the error to are more meaningful message along with how to
proceed, ie. try again.