Detect request EOF errors
If a client is not connected to the internet or Lunar VPN they receive two
implementation detail errors.

    Error: Post https://release-manager.dev.lunarway.com/release: dial tcp: lookup release-manager.dev.lunarway.com: no such host
    Error: Post https://release-manager.dev.lunarway.com/release: EOF

One indicates no DNS lookup is possible. The other that we cannot connect to the
found IP (if it is inside our VPN).

This change detects these two error cases in the hamctl CLI and will instead
give a more descriptive error.

    Error: could not connect to the release-manager server. Are you connected to the internet and, if required, a VPN?