Remove error log on unknown email (#132)
If an unknown email is used to release a resource, we log the error in the slack
package and return an error. The error is then logged in the HTTP handlers and a
500 error is returned to the daemon.

This results in the daemon also logs an error.

This changes ignores the email errors in the handlers and returns a 200 to the
daemon. This will result in a single error log on unknown emails.