Remove Git operations when notifying comitter
This change removes the need for locating the artifact spec when handling daemon
webhooks. We can derive all the necessary information from the pod annotations
and pod event instead.

This reduces heavy load on the server if multiple pods are in CrashLoopBackoff
as we no longer do any git operations nor expensive cp operations for every pod
start/fail event.