Add context logging
This change adds two context logging helpers. AddContext will enrich a context
instance with fields that can later be extracted with WithContext. The logging
instance returned by the latter will log the extracted fields with any logged
line.

The trace middleware enriches the context with the request id and all other
functions and methods are updated to always log with a context logger so that
the request id can be used as a reference for all related lines.