Make muteable endpoints asyncronous (#140)
In order to provide a better developer experience on large config repositories
(>10000 commits) where release and other mutating commands are getting slow, ie.
10s of seconds, we are going to make it fully asynchronous.

This means that we only do some initial input validation before publishing an
event to RabbitMQ and returning a response to the client. The server will then
consume these events one by one.

This will reduce the response time of mutating commands and further ensure that
their operation is completed as soon as possible instead of letting the client
in charge for retrying. It will also reduce the number of conflicts when
multiple clients are trying to write to the cluster config repository. Before
this, clients had to inspect the response in this case and retry the attempts,
now it will be handled by only handling a single event at a time.

# Implementation details

## AMQP publisher and consumer

Package amqp implements an AMQP Worker capable of handling connection failures
and other fatal errors on the AMQP connection. It exposes two methods providing
the general functionality of the AMQP protocol: StartConsumer and Publish.

The former is a blocking method that starts consuming messages on the configured
queue with provided handlers for the different event types. The latter publishes
a message on the configured exchange.

## Flows

Release, promote and rollback flows are updated to only do required input
validation, ie. check for available artifacts to the environments. They will
then publish a message on AMQP that is handled by the Exec* variants of the
methods that performs the actual manipulation of the config repository.

# Open TODOs

There are a number of improvements that could be made that I've only noted as
TODOs right now. They'll be fixed in upcoming PRs to keep things simple and get
this into master.

## Non-async flow

A flag to replace the publisher/consumer with an in-memory queue to ease local
development and deployment.

## Nack events with dead letter queue

Events are not nacked/acked in case of event handler errors. With a low prefetch
(default 1) this means that event handling is blocked after the first failed
event.

To fix this we can either just drop the events (nack and no requeue) and inform
the user of what happened. Another option is to add a dead letter queue and push
events over there. This does however require a mechanism for persisting them
from there and further redelivering/dropping them.

I'm most in favour on the former one. Errors could be like invalid configuration
(no config for environment) etc and these bugs should be handled in the sync
validation step instead.

## Cancel consumer before closing it

The Close method of the consumer type in package amqp closes its AMQP channel
but does not cancel it before hand. This can potentially lead to dropping
inflight events as the channel is closed before all events are consumed.

We can handle this by adding a wait group or something like that guarding the
Start method that is the consumer. With this we can cancel the channel, wait for
the consumer to stop (no more messages) and then close the channel.

## Publish confirms

AMQP provides a mechanism for guaranteeing published messages to have arrived on
consumer queues. This is not handled for now and can potentially drop events if
the remote AMQP server fails in forwarding the message to queues.

## Environment up to date messages

Going async means that the git status identifying that we have nothing new to
release is handled in the async execution path. There is no message to the
releaser for now except a log message so eventually we want to push this message
out in Slack. A even better solution would be to validate the released artifact
ID against the current one in the target environment and thus provide this
message in the synchronous flow.