Add in-memory broker
This change set adds an in-memory broker implementation as a replacement for the
AMQP broker. This makes is simpler to run the server locally or in environments
where percistence of the queue is not necessary.

Package amqp is moved into package broker along with the memory broker and the
new broker package exports the interfaces required by a broker.

A new flag broker-type is added to configure what implementation is to be used
and the Make target server is replaced by server-memory and server-rabbitmq
configured respectively with the in-memory and amqp brokers.

A small change to the contract between the publisher and consumer is added as
well. Event implementations are now required to implement both an Unmarshal
(former Body) method along with a Marshal method. This makes them selfcontained
and removes the odd encoding/json package dependency of package amqp.

The handler registration is also moved to the StartConsumer method instead of
being required at construction time. This makes for a simpler implementation of
brokers and also makes more sense in the domain of handlers being coupled with
consumers and not brokers in general.