Fail fast on nothing to release
If a release command specifies what is already running, e.g. two releases from
the same branch without new artifacts, no feedback is given to the client. A log
message in the asynchronous flow is the only indication what there was nothing
to do.

This change set adds a validation of the artifact ids as to detect this up front
and return a message to the client as fast as possible.