Add branch-restriction policy
This change set includes a branch restriction policy that will limit what
branches are possible to release to what environments.

It can be added with hamctl like the existing auto-release policy

    hamctl policy apply branch-restriction --env prod --matcher '^master$'

The matcher is parsed as a regular expression and must be parsable as such.
Validation is done when applying the policy and when ever it is used.

The mutating release flows are updated to verify if the artifact that is to be
released is allowed and auto release policies are verified against any branch
restrictions before being applied.