Always log errors in server HTTP handlers
Currently some of the error cases are identified and custom messages are both
logged and returned to clients. Some of these logs does not include to original
error which effectively removes any context that might be relevant.

This stems from the assumption that all errors are only observable through
exactly one code path, but with the introduction of synchronous detection of up
to date environments the artifact.ErrFileNotFound can now be found in two code
paths, ie. looking up source and current artifact.

This change set adds the origin errors to all log messages in the HTTP handlers
and thus adding the context of each one of them.