Ignore missing artifacts from existing release
When testing if the current released artifact is the same as a new release we
might hit an artifact.ErrFileNotFound error if there has never been a release to
the target environment.

This change set ensures that this error is ignored in these checks to avoid
effectively blocking releases to new environments.

It also adds a check to the release artifact flow that was missed in
aff012fd6a433664956876a5adfa640cfcb0a9ac.