Inspect added files in GitHub webhooks (#160)
If a GitHub webhook is based of the first artifact for a service or branch it
will only contain added files. Currently we only inspect modified files and thus
we will fail to auto-release the very first artifact from a branch matching the
policy.

This change merged added and modified files to a single slice and detects the
branch name from these. This will ensure that we find a branch name also for the
very first artifacts on a branch.