Add publish confirm mode for AMQP deliveries
This change set implements publish confirmation in the AMQP broker. It will
ensure that the AMQP server confirms all publishes and that the Publish method
does not return before this confirmation is received.

The timeout between publishes can be configured with flag
amqp-republish-timeouts and defaults to 1 second. During republishes contexts
cancellation and shutdowns are respected and will result in context.Cancelled
and broker.ErrBrokerClosed errors respectively.