Fix branch restriction for same env as auto release (#165)
If we try to apply a branch restriction policy to a branch that matches an auto
release policy but for another environment the apply fails as if they conflict.

This is due to a missing environment comparision when checking against the auto
release policies. This comparison is added in this change set.