Support global only policies
If only global policies are applied for a service the policy.Service.Get method
returns ErrNotFound instead of the global policies. This leads to listing
policies indicates no policies along with any release restrictions are not
enforced.

This change set detects this case and returns the globally applied policies in
this case.