Compare artifact releases correctly (#172)
When releasing a specific artifact the sync flow validates whether there will be
anything to release by comparing the released artifact id to the request one.
The comparison was done on the artifacts source Git sha instead of the HEAD SHA.
This is a bug introduced in 0a99b155c382e517b1785fdeb5b0f773dff632af (#161)
where the check was added but the origin of the environment spec was wrong.

When handling release.artifactId events the server did also not checkout the
hash of the artifact and thus release the latest artifact for the service
instead of the specified one. This bug was introduced during the async flow
change b1c9fe82f11e11cd2abb1e084ead74dc2f0a8e9c (#140) where the checkout was
only kept in the sync section of the flow.