Refactor artifact storage into separate package (#178)
This change set contains changes to package flow to extract the artifact storage
layer into an interface. The storage is still Git for now, but the interface is
not implemented by package git instead. This will eventually be replaced with
AWS S3 blob storage and with the new flow.ArtifactReadStorage interface it is
clear what a storage layer needs to provide.

This will also make it simpler to implement in-memory storages for testing the
flows without touching external mechanisms.

The heart of the change is as mentioned the flow.ArtifactReadStorage interface.

    type ArtifactReadStorage interface {
      // ArtifactExists returns whether an artifact with id artifactID is available.
      ArtifactExists(ctx context.Context, artifactID string) (bool, error)

      // ArtifactSpecification returns the artifact specification for a given
      // service and artifact ID.
      ArtifactSpecification(ctx context.Context, service, artifactID string) (artifact.Spec, error)

      // ArtifactPaths returns file system paths for the artifact specification
      // (specPath) and yaml resources directory (resourcesPath) available on the
      // file system for copying to releases. The returned close function is
      // responsible for clean up of the persisted files.
      ArtifactPaths(ctx context.Context, service, environment, branch, artifactID string) (specPath, resourcesPath string, close func(context.Context), err error)

      // LatestArtifactSpecification returns the latest artifact specification for a
      // given service and branch.
      LatestArtifactSpecification(ctx context.Context, service, branch string) (artifact.Spec, error)

      // LatestArtifactPaths returns file system paths for the artifact
      // specification (specPath) and yaml resources directory (resourcesPath)
      // available on the file system for copying to releases of the latest artifact
      // for provided service and branch. The returned close function is responsible
      // for clean up of the persisted files.
      LatestArtifactPaths(ctx context.Context, service, environment, branch string) (specPath, resourcesPath string, close func(context.Context), err error)

      // ArtifactSpecifications returns a list of n newest artifact specifications
      // for service. They should be ordered by newest first.
      ArtifactSpecifications(ctx context.Context, service string, n int) ([]artifact.Spec, error)
    }

It allows the flow package to check for the existence of artifacts along with
fetching specifications for a specific artifact of a n artifacts.

The ArtifactPaths and LatestArtifactPaths methods are the methods providing
access to the actual resources of an artifact. Currently these are paths to the
file system, but I think we should come up with something better later on. Maybe
it will become clear with the S3 storage implementation what makes sense.

Bug fix in promote commit messages During the change I found a bug in the commit
messages of promote operations. The are currently resulting in the commit
message [env/service] release by bso@lunar.app. Note the missing artifact ID.
This was introduced in b1c9fe8 (#140).