Fix bug in completion annotations
Currently if a hamctl release command is issued without an 'env' flag the
command is accepted and pushed to the server. The flag is marked required but
due to a bug in completion.FlagAnnotation the required annotation is dropped.

This change set ensures to not change existing annotations on the flag before
setting the bash completion annotation