Remove git artifact storage
This change removes all code and docs related to Git stored artifacts.

This is a breaking change for clients as artifact's flags ssh-private-key and
config-repo are removed and http-base-url and http-auth-token are marked
required.

It will also break any usage of artifacts in the fallback storage mechanism ie.
if artifacts are not found in S3 we return an error.