Add missing error checks
If the ReleaseArtifactID command fails in hamctl release no error is written to the screen.

If we cannot create a new artifact on S3 events we also do nothing to record the error.

This change adds error checks to both these places.

Found with go get -u github.com/kisielk/errcheck

    errcheck ./...