Fix bug in commit message parsing for flux commits (#207)
Flux only publishes the first line of commit messages when syncing. After the
introduction of multiline commit messages in the release-manager and the shared
handling of commit parsing in package commitinfo we broke the notifications for
running pods.

This change fixes the byg by handling commits of type "auto release" separatly
when parsing the intent.

The regex used to parse the commit info is also updated to allow multi word
"types", e.g. "auto release".

Lastly logs of flux's event.Event type is expanded to all fields instead of its
String implementation. Currently a prettyfied string is shown in the logs
instead of the complete contents.