Remove credentials requirement from hamctl cmds
Currently hamctl commands "version" and "completion" will fail if valid Git
credentials are not found. In practice this becomes a problem when evaluating
completions into a file as the file will only contain the error output.

This change introduces a persistent user-email flag for controlling who the
author of the command execution is. It will default to Git configured user.email
as before. The value is ignored for above mentioned commands.

If the Git user.email value is empty for a command requiring it an error is
returned as before.