Fix path traversal in policy and status endpoints
Currently providing service names like ../ enables path traversal on policy
lookups along with status checks.

This is due to the service paramater is not sanitized before used with os.Open
and os.OpenFile.

This change set adds sanitization ensuring that paths cannot point to parents
directories of the action performed. The dependency
github.com/cyphar/filepath-securejoin is added to handle this sanitization.

A GitService interface is introduced in the package flow to allow for testing
the releasePath method ensuring that this works properly.

The vulnerability is reported by GitHub Code scanning in the following tickets.

    https://github.com/lunarway/release-manager/security/code-scanning/1
    https://github.com/lunarway/release-manager/security/code-scanning/2