Fix arbitrary file write in zip extract
Currently a zip archive with file paths like ../ can result in file overwrites on the host system outside the destination directory.

This change set adds sanitization to the zip archive's file names, ensuring that paths cannot point to parent directories of the destination path.

The vulnerability is reported by GitHub Code scanning in the following ticket.

    https://github.com/lunarway/release-manager/security/code-scanning/3