Ignore nothing to release errors on new artifacts (#212)
Currently if a new artifact command results in nothing to release in the auto
release step the event is failed and re-queued. As this is an expected error state
in case of events being handled multiple times or if another release command has
raced it, we can just ignore the error.

This change set adds the error to the ignored errors in the auto release error
handling.