Remove weaveworks/flux dependency (#219)
Currently we depend on the github.com/weaveworks/flux module to parse flux
websocket events. We only rely on a couple of fields in the data model but
depend on all transitive dependencies of the flux project.

Further more the flux module has moved into github.com/fluxcd/flux after being
adopted by CNCF. A change in the new module structure does not allow us to
import the used packages anymore.

This change removes the dependency by copying the used models into this project.
It has little impact on readability but simplifies the dependency tree.