Fix missing releases in describe and rollback (#222)
Currently some releases are not reported when using hamctl describe release and
hamctl rollback.

This happens as on each matched release we checkout the repository on the found
hash and inspect some files. The HEAD is moved on each of these checkouts which
makes our search offset wrong.

This change fixes the bug by resetting the branch to the original HEAD after
each checkout ensuring our search offset is aligned.

It is not possible to test the DescribeRelease method in a simple way as it
depends on multiple git details and the file system. Because of this the change
set includes a unit test that only validates the core git logic of the flow in a
controlled manner. Eventually we should refactor the flows to be more testable
and move over to testing the actual implementation.