Add branch filter to describe artifacts
This change adds a branch filter to /describe/artifacts that can be used to only
get artifacts from a specific branch.

To test the otherwise untested ArtifactSpecifications method a new dependency is
added for faking S3 storage. This makes it trivial to seed a bucket with test
data and use that in tests. The added test TestService_ArtifactSpecifications
covers existing functionality along with the new branch filter.

To make the artifactreadstoreage.Service testable the AWS clients and session is
extracted out of the constructor and instead injected in start.go on the server.