Add branch filter to describe artifacts (#242)
This change adds a branch filter to /describe/artifacts that can be used to only
get artifacts from a specific branch.

To test the otherwise untested ArtifactSpecifications method a new dependency is
added for faking S3 storage. This makes it trivial to seed a bucket with test
data and use that in tests. The added test TestService_ArtifactSpecifications
covers existing functionality along with the new branch filter.

To make the artifactreadstoreage.Service testable the AWS clients and session is
extracted out of the constructor and instead injected in start.go on the server.

The existing S3 testing setup is removed. It was only enabled locally if we had
S3 credentials setup so provided little value. The new S3 backend is now used
and tests are run in CI.