Align hamctl template usage and decople data model (#246)
Currently we have three ways of writing output to the screen in hamctl. 1)
fmt.Print variants for simple line outputs, 2) fmt.Print variants for complex
outputs and 3) Go templates.

1) is ok for the most simple outputs and are left as is. 2) is hard to read and
modify and are updated to use templates instead.

All the temlpates are updated to use a dedicated data model to avoid coupling
HTTP DTOs to the outputs which is error prone and hard to catch at compile time.