Add tests of hamctl status template
Add unit tests of the status template to ensure it renderes correctly. It did
not, so there are several small changes to both the look and layout.

Along the way a bug was found in the commonly used format string so this is
extracted out to a shared template variable to control how dates are presented.