Add dynamic Grafana annotation flag (#257)
Currently Grafana annotations are hard coded with dedicated flags for each
supported environment. This makes the code hard coupled and cumbersome to expand
with new environments.

This change adds a new grafana-annnotations flag that accepts a comma
separated list of <env>=<api-key>=<url> values. This makes it possible to add
as many environments as needed without making changes to the code base.

The grafanaOptions struct impements the pflag.Value and pflag.SliceValue
interfaces which makes pflag handle multiple apperances of the flag correctly.
This means the following flag settings are all valid and will result in 4
configured environments.

    server start \
      --grafana-annotations dev=${GRAFANA_API_KEY}=${GRAFANA_URL} \
      --grafana-annotations staging=${GRAFANA_API_KEY}=${GRAFANA_URL} \
      --grafana-annotations prod=${GRAFANA_API_KEY}=${GRAFANA_URL},platform=${GRAFANA_API_KEY}=${GRAFANA_URL}