Do not log unknown grafana environment as errors (#259)
Currently if the server tries to annotate a release in an environment where
Grafana credentials are not configured it logs an error.

As this is not an error of the server this change changes behaviour to log an
info line in this case. This should be enough for operators to understand why
Grafana annotations are missing and not create error logs if this is
intentional, ie. no Grafana environment available.